GEODE-9639: Removed remaining deprecations that can be removed.

+++ cppcache/src/ThinClientPoolDM.cpp
@@ -198,7 +198,7 @@ ThinClientPoolDM::ThinClientPoolDM(const char* name,
   if (!props.isEndpointShufflingDisabled()) {
     if (!m_attrs->m_initServList.empty()) {
       RandGen randgen;
-      m_server = randgen(static_cast<uint32_t>(m_attrs->m_initServList.size()));
+      m_server = randgen(m_attrs->m_initServList.size());
     }
   }
   if (m_attrs->getPRSingleHopEnabled()) {
